requirements.txt
faiss-cpu
google-search-results
gradio
langchain
langchain_experimental
openai
tiktoken
